Códigos del capitulo 3 completos
