+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+    "files.associations": {
+        "stdlib.h": "c"
+    }
+}
