--- src/Cap3/listing3.2/listing3.2.c
@@ -1,7 +1,0 @@
-#include <stdlib.h>
-int main()
-{
-    int return_value;
-    return_value = system("ls -l /");
-    return return_value;
-}
