+++ src/Cap3/listing3.2/system.c
@@ -1,0 +1,7 @@
+#include <stdlib.h>
+int main()
+{
+    int return_value;
+    return_value = system("ls -l /");
+    return return_value;
+}
