+++ src/Cap3/listing3.7/sigchld.c
@@ -19,6 +19,23 @@ int main()
     sigchld_action.sa_handler = &clean_up_child_process;
     sigaction(SIGCHLD, &sigchld_action, NULL);
     /* Now do things, including forking a child process. */
-    /* ... */
+    // Creamos el proceso hijo
+    pid_t child_pid;
+    // Clonamos el proceso
+    child_pid = fork();
+    if (child_pid != 0)
+    {
+        printf("Id proceso padre: %d\n", getpid());
+        printf("Id proceso hijo: %d\n", child_pid);
+        // El proceso padre duerme por 100s
+        sleep(100);
+    }
+    else
+    {
+        printf("Id proceso hijo: %d\n", child_pid);
+        // El proceso hijo duerme por 10 s
+        sleep(10);
+    }
+
     return 0;
 }
